Final Release
Plugin is finally out of the Beta Era. After months of development, I
believe this is something that is good enough for use with any server.
Backwards and forwards compatibility will now become the primary focus,
and new features  won't be implemented as often. It's an exciting time
for this API to shine!