ADDED SUPPORT FOR MENDING! :D
Really excited because I was thinking it was impossible. I even made a
thread on Spigot and everyone agreed. BUT Vanilla Minecraft does try to
increase the Durability of Unbreakable Items with Mending, but fails and
gives the xp back to the player, which I can then detect with the
PlayerExpChangeEvent and increase teh fake durability. That being said,
this means that I can't a modifier, because Fake Durability on non
unbreakable items will still function with Vanilla's Code.