MAJOR UPDATES
I'll push more often, I promise :P

RUNTIME TRANSFORMER ONLY WORKS ON LINUX. Changes have to be made to work
on Windows, might seem simple at first but it's much deeper.

Also switched to Maven and IntelliJ. Change sucks, but it was needed. If
you have to compile, you'll have to download a Spigot jar and change the
pom, as I need NMS which is not available in Online Repositories or
attainable by running BuildTools.

THIS UPDATE WAS PAINFUL.