THE DEFINITIVE RESOURCE PACK SOLUTION
Truly a work of art (and bodging - especially that). It starts by creating the resource pack file, sets its date to a specific one, then generates the md5 of the file. Since the file and its contents have a specific date, we can detect if the resource pack has any changes by checking the saved md5 in the config. Then, if changes are detected, it is uploaded to Minepack.net, which might already have the file, so we'll use that if it's already there and generate the SHA-1 and the link ourselves. Otherwise, it's uploaded and the player is sent the link when he joins.